[ridho] - create policy for update & delete PlaylistPolicy

+++ app/Http/Controllers/Screencast/PlaylistController.php
@@ -38,12 +38,17 @@ class PlaylistController extends Controller
 
     public function edit(Playlist $playlist)
     {
+        // if ($playlist->user_id !== Auth::user()->id) {
+        //     return redirect(route('table.playlists'));
+        // }
+        $this->authorize(['update', $playlist]);
         $tags = Tag::get();
         return view('playlists.edit', compact([ 'playlist', 'tags']));
     }
 
     public function update(PlaylistRequest $request, Playlist $playlist)
     {
+        $this->authorize(['update', $playlist]);
         if ($request->thumbnail) {
             Storage::delete($playlist->thumbnail);
             $thumbnail = $request->file('thumbnail')->store('images/playlists');
@@ -70,6 +75,7 @@ class PlaylistController extends Controller
 
     public function destroy(Playlist $playlist)
     {
+        $this->authorize(['delete', $playlist]);
         Storage::delete($playlist->thumbnail);
         $playlist->tags()->detach();
         $playlist->delete();
